feat(ui): enhance settings menu with submenu and dynamic visibility
- Replaced static `value="false"` in `v-list-group` with `v-model` bound to
`subMenuState` for dynamic submenu state handling.

- Introduced computed properties for namespaced instance and namespaces count.

- Added dynamic submenu items to the Settings menu, including "Profile,"
"Namespace," "Private Keys," "Tags," and "Billing," with conditional
visibility based on environment variables.

- Initialized `subMenuState` as a reactive object for managing
expanded/collapsed states of submenus.

- Removed unused `SettingDrawer` in `Settings.vue` and adjusted its snapshot
test accordingly.
