+++ ui/src/views/Settings.vue
@@ -1,13 +1,8 @@
 <template>
-  <SettingDrawer v-model="showNavigationDrawer" />
-
   <div>
     <router-view />
   </div>
 </template>
 
 <script setup lang="ts">
-import SettingDrawer from "@/components/Setting/SettingDrawer.vue";
-
-const showNavigationDrawer = defineModel();
 </script>
